Fully implemented security

+++ demo/src/main/java/com/orienteering/rest/demo/security/JwtTokenProvider.java
@@ -41,7 +41,7 @@ public class JwtTokenProvider {
         Date expiryDate = new Date(now.getTime() +jwtExperationInMs);
 
         return Jwts.builder()
-                .setSubject(Integer.toString(userPrincipal.getId()))
+                .setSubject(Long.toString(userPrincipal.getId()))
                 .setIssuedAt(new Date())
                 .setExpiration(expiryDate)
                 .signWith(SignatureAlgorithm.HS512, jwtSecret)
